wrapper refac pt2
